Rename examples, tests directories to use singular form.
This is for consistency with prior art.

Also rename ./test/integration package from tests to integration.
It makes no difference what the package is called, since it's only
used as a container for tests. But it's more common to name packages
after the directory they're in, so do that.

Resolves #671.
