+++ example/basicauth/main.go
