+++ test/README.md
@@ -37,7 +37,7 @@ These tests require a GitHub user (username and password), and also that a
 attendant Client ID and Client Secret) be available. Then, to execute just the
 Authorization tests:
 
-    GITHUB_USERNAME='<GH_USERNAME>' GITHUB_PASSWORD='<GH_PASSWORD>' GITHUB_CLIENT_ID='<CLIENT_ID>' GITHUB_CLIENT_SECRET='<CLIENT_SECRET>' go test -v -tags=integration --run=Authorizations ./integration
+    GITHUB_USERNAME='<GH_USERNAME>' GITHUB_PASSWORD='<GH_PASSWORD>' GITHUB_CLIENT_ID='<CLIENT_ID>' GITHUB_CLIENT_SECRET='<CLIENT_SECRET>' go test -v -tags=integration -run=Authorizations ./integration
 
 If some or all of these environment variables are not available, certain of the
 Authorization integration tests will be skipped.
