+++ test/integration/doc.go
@@ -3,9 +3,9 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package tests contains integration tests.
+// Package integration contains integration tests.
 //
 // These tests call the live GitHub API, and therefore require a little more
-// setup to run. See https://github.com/google/go-github/tree/master/tests#integration
+// setup to run. See https://github.com/google/go-github/tree/master/test#integration
 // for more information.
-package tests
+package integration
