+++ test/integration/issues_test.go
@@ -5,7 +5,7 @@
 
 // +build integration
 
-package tests
+package integration
 
 import "testing"
